refactor(form-field): Update stories
