+++ modules/common/css/lib/forms.scss
@@ -62,6 +62,14 @@
   min-height: $wdc-spacing-s + $wdc-spacing-xxxs;
 
   input {
+    border-radius: 2px;
+    width: $wdc-spacing-m;
+    height: $wdc-spacing-m;
+    margin: 0;
+    margin-top: -3px;
+    margin-left: -3px;
+    position: absolute;
+    opacity: 0;
     &:hover + label::after {
       box-shadow: 0 0 0 7px $wdc-color-soap-200;
     }
@@ -73,8 +81,8 @@
       color: $wdc-color-black-pepper-300;
       font-family: 'Roboto', 'Helvetica Neue', 'Helvetica', Arial, sans-serif;
       position: relative;
-      left: $wdc-spacing-xs;
-      padding-right: $wdc-spacing-xs;
+      left: $wdc-spacing-xs + 18px;
+      padding-right: $wdc-spacing-xs + 18px;
       cursor: pointer;
 
       // Input styling
@@ -99,7 +107,7 @@
         display: inline-block;
         position: absolute;
         border-radius: 16px;
-        top: -1px;
+        margin-top: 1px;
         left: -30px;
         width: 18px;
         height: 18px;
