+++ modules/form-field/css/stories/stories_Radio.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import withReadme from 'storybook-readme/with-readme';
-import README from '../README.md';
+import README from '../../../radio/css/README.md';
 import '../index.scss';
 import '@workday/canvas-kit-css-radio/index.scss';
