+++ modules/form-field/css/stories/stories_Select.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import withReadme from 'storybook-readme/with-readme';
-import README from '../README.md';
+import README from '../../../select/css/README.md';
 import '../index.scss';
 import '@workday/canvas-kit-css-select/index.scss';
