+++ modules/form-field/css/stories/stories_TextArea.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import withReadme from 'storybook-readme/with-readme';
-import README from '../README.md';
+import README from '../../../text-area/css/README.md';
 import '../index.scss';
 import '@workday/canvas-kit-css-text-area/index.scss';
